requirements.txt
pytest==7.4.3
requests==2.31.0
allure-pytest==2.13.2
allure-python-commons==2.13.2

